fix(sentry-cli): fix download URLs
Unfortunately, Sentry has inconsistent download URLs, e.g. same arch has different name across os'es. As far as I know it is not possible to configure all of this with non-WASM plugins. This commit is a best-effort, realizing arm for MacOS and x64 for Linux/Windows (along with some other combinations)